H3DAPI warning fixes (vs2010 64-bit)
Summary:
Additional fixes for warnings that only show up when doing 64-bit builds (primarily because size_t and int are no longer the same size in 64-bit visual studio)

Related task: T712

Test Plan: Will run unit tests which should open not only all the directly edited nodes but also many of the nodes that might be indirectly affected by any of the changes.

Reviewers: markus

Reviewed By: markus

Subscribers: michaels, markus

Tags: #h3dapi_2.x

Differential Revision: https://www.h3dapi.org:4000/D33

+++ H3DAPI/include/H3D/H3DDisplayListObject.h
@@ -185,7 +185,7 @@ namespace H3D {
             
       /// Returns the number of input fields that has generated an event
       /// to this field since the last call to the update() function.
-      inline unsigned int nrPendingEvents() {
+      inline size_t nrPendingEvents() {
         return event_fields.size();
       }
